refactor extract deactivateAction
